requirements.txt
dm-haiku
jaxopt
optax
matplotlib
numpy
scikit_learn
scipy
tqdm
pandas
submitit